refactor: split logic into container and component
