+++ my-app/src/MockData/data.tsx
@@ -8,6 +8,7 @@ export default [
     "picture": "https://photos.zillowstatic.com/fp/429e4774150e3e617f7ff66767ec8b8a-p_e.jpg",
     "houseType": "detached house",
     "email": "glassdickerson@ziggles.com",
+    "dateOfPublish": "1990-02-01",
     "phone": "+1 (885) 443-3753",
     "address": "823 Kay Court, Blanco, Ohio, 6067",
     "about": "Deserunt pariatur quis quis dolore duis exercitation quis laboris. Esse nisi voluptate cupidatat irure do elit eu. Proident incididunt fugiat exercitation occaecat qui dolore ad commodo occaecat amet id aute. Quis nisi ea consequat eu.\r\n"
@@ -20,6 +21,7 @@ export default [
     "picture": "https://photos.zillowstatic.com/fp/6c170661eb73c9f87b404611ccbc9598-p_e.jpg",
     "houseType": "townhouse",
     "email": "glassdickerson@uberlux.com",
+    "dateOfPublish": "1990-02-01",
     "phone": "+1 (929) 545-2778",
     "address": "645 Richardson Street, Escondida, Virginia, 8798",
     "about": "Sunt aliqua minim nostrud magna tempor dolor. Mollit magna minim amet id id. Culpa consequat laborum veniam veniam. Eiusmod elit ad non consequat.\r\n"
@@ -32,6 +34,7 @@ export default [
     "picture": "https://photos.zillowstatic.com/fp/60a5d08dc2439555171ce2e756d6a8dc-p_e.jpg",
     "houseType": "detached house",
     "email": "glassdickerson@megall.com",
+    "dateOfPublish": "1990-02-01",
     "phone": "+1 (803) 598-3383",
     "address": "313 Tennis Court, Cutter, Florida, 6294",
     "about": "Consequat quis esse non sit et est. Ea tempor minim ex pariatur consectetur in quis. Aliquip nulla aliquip aliqua exercitation aliquip reprehenderit nostrud sit velit sit consequat cupidatat. Veniam do deserunt laborum cupidatat ut. Lorem id commodo et exercitation ullamco duis occaecat Lorem pariatur tempor proident.\r\n"
@@ -44,6 +47,7 @@ export default [
     "picture": "https://photos.zillowstatic.com/fp/50f67dfa622f0e86409ff67bfbb6f690-p_e.jpg",
     "houseType": "detached house",
     "email": "glassdickerson@unisure.com",
+    "dateOfPublish": "1990-02-01",
     "phone": "+1 (920) 533-2160",
     "address": "139 Duffield Street, Hoagland, West Virginia, 7811",
     "about": "Dolore pariatur velit qui ut dolor amet voluptate culpa enim ex. Mollit ut veniam laborum enim ad officia dolor. Duis ut et do voluptate fugiat deserunt magna voluptate mollit. Deserunt adipisicing irure tempor consequat non.\r\n"
@@ -56,6 +60,7 @@ export default [
     "picture": "https://photos.zillowstatic.com/fp/145b38e7bd3cdcb0416b6c599de564a4-p_e.jpg",
     "houseType": "townHouse",
     "email": "glassdickerson@proflex.com",
+    "dateOfPublish": "1990-02-01",
     "phone": "+1 (818) 429-3849",
     "address": "292 Autumn Avenue, Lorraine, Pennsylvania, 8194",
     "about": "Mollit nulla velit labore officia irure veniam anim laboris ut irure culpa velit. Ea magna consequat aliqua culpa sit labore anim aute id eu sint in occaecat. Excepteur anim nisi voluptate ad anim aliquip enim ullamco cillum amet nulla nostrud eu. Incididunt ad esse exercitation irure minim id Lorem aliquip incididunt. Reprehenderit reprehenderit esse adipisicing in id ipsum dolor quis voluptate.\r\n"
